Mitigate inconsistent behaviour on slow systems
The onDidChangeActiveTerminal event does not appear to play well with the timing of creating new terminals. This alternative implementation removes the event dependency, and rather utilises simple logic and the known state of window.terminals[]. Also added delays to allow proper finalisation of terminal command execution - without these, it executes before the terminals are in an acceptable ready state. These can be parametrised. Before the change, it failed to restore terminals in about 4/5 attempts. After the change, it succeeds every time.